Imoproved sync repository sql script execution
